Add audio transcription and text-to-speech support to API
- Introduced new `/v1/audio/transcriptions` endpoint for audio transcription using Whisper and Parakeet models.
- Added `/v1/audio/speech` endpoint for text-to-speech functionality.
- Updated API documentation to reflect new audio capabilities.
- Enhanced build script to include necessary audio dependencies and ensure proper bundling.
- Implemented runtime checks for audio library availability in the application.
- Improved model management to reset statuses and update memory requirements based on actual file sizes.
- Enhanced admin interface to categorize and display audio models effectively.

+++ src/mlx_gui/models.py
@@ -55,7 +55,7 @@ class Model(Base):
     version = Column(String)
     model_type = Column(String, nullable=False)
     huggingface_id = Column(String)
-    memory_required_gb = Column(Integer, nullable=False)
+    memory_required_gb = Column(Float, nullable=False)
     status = Column(String, nullable=False, default=ModelStatus.UNLOADED.value)
     created_at = Column(DateTime, default=datetime.utcnow)
     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
